Add widget cache for plugin components
